put win32,posix,standard all in env.c, minor fixes, and astyled
