+++ src/md5.h
@@ -27,7 +27,7 @@
 
   This code implements the MD5 Algorithm defined in RFC 1321, whose
   text is available at
-	http://www.ietf.org/rfc/rfc1321.txt
+    http://www.ietf.org/rfc/rfc1321.txt
   The code is derived from the text of the RFC, including the test suite
   (section A.5) but excluding the rest of Appendix A.  It does not include
   any code or documentation that is identified in the RFC as being
@@ -38,12 +38,12 @@
   that follows (in reverse chronological order):
 
   2002-04-13 lpd Removed support for non-ANSI compilers; removed
-	references to Ghostscript; clarified derivation from RFC 1321;
-	now handles byte order either statically or dynamically.
+    references to Ghostscript; clarified derivation from RFC 1321;
+    now handles byte order either statically or dynamically.
   1999-11-04 lpd Edited comments slightly for automatic TOC extraction.
   1999-10-18 lpd Fixed typo in header comment (ansi2knr rather than md5);
-	added conditionalization for C++ compilation from Martin
-	Purschke <purschke@bnl.gov>.
+    added conditionalization for C++ compilation from Martin
+    Purschke <purschke@bnl.gov>.
   1999-05-03 lpd Original version.
  */
 
@@ -66,9 +66,9 @@ typedef unsigned int mongo_md5_word_t; /* 32-bit word */
 
 /* Define the state of the MD5 Algorithm. */
 typedef struct mongo_md5_state_s {
-    mongo_md5_word_t count[2];	/* message length in bits, lsw first */
-    mongo_md5_word_t abcd[4];		/* digest buffer */
-    mongo_md5_byte_t buf[64];		/* accumulate block */
+    mongo_md5_word_t count[2];  /* message length in bits, lsw first */
+    mongo_md5_word_t abcd[4];       /* digest buffer */
+    mongo_md5_byte_t buf[64];       /* accumulate block */
 } mongo_md5_state_t;
 
 #ifdef __cplusplus
@@ -76,14 +76,14 @@ extern "C"
 {
 #endif
 
-/* Initialize the algorithm. */
-MONGO_EXPORT void mongo_md5_init(mongo_md5_state_t *pms);
+    /* Initialize the algorithm. */
+    MONGO_EXPORT void mongo_md5_init(mongo_md5_state_t *pms);
 
-/* Append a string to the message. */
-MONGO_EXPORT void mongo_md5_append(mongo_md5_state_t *pms, const mongo_md5_byte_t *data, int nbytes);
+    /* Append a string to the message. */
+    MONGO_EXPORT void mongo_md5_append(mongo_md5_state_t *pms, const mongo_md5_byte_t *data, int nbytes);
 
-/* Finish the message and return the digest. */
-MONGO_EXPORT void mongo_md5_finish(mongo_md5_state_t *pms, mongo_md5_byte_t digest[16]);
+    /* Finish the message and return the digest. */
+    MONGO_EXPORT void mongo_md5_finish(mongo_md5_state_t *pms, mongo_md5_byte_t digest[16]);
 
 #ifdef __cplusplus
 }  /* end extern "C" */
