+++ src/mongo.c
@@ -19,10 +19,7 @@
 #include "md5.h"
 #include "env.h"
 
-#include <stdlib.h>
-#include <stdio.h>
 #include <string.h>
-#include <stdlib.h>
 
 MONGO_EXPORT mongo* mongo_create( void ) {
     return (mongo*)bson_malloc(sizeof(mongo));
@@ -30,7 +27,7 @@ MONGO_EXPORT mongo* mongo_create( void ) {
 
 
 MONGO_EXPORT void mongo_dispose(mongo* conn) {
-    free(conn);
+    bson_free(conn);
 }
 
 MONGO_EXPORT int mongo_get_err(mongo* conn) {
@@ -100,7 +97,7 @@ MONGO_EXPORT mongo_cursor* mongo_cursor_create( void ) {
 
 
 MONGO_EXPORT void mongo_cursor_dispose(mongo_cursor* cursor) {
-    free(cursor);
+    bson_free(cursor);
 }
 
 
@@ -114,7 +111,7 @@ MONGO_EXPORT const char*  mongo_get_server_err_string(mongo* conn) {
 }
 
 MONGO_EXPORT void __mongo_set_error( mongo *conn, mongo_error_t err, const char *str,
-                               int errcode ) {
+                                     int errcode ) {
     int errstr_size, str_size;
 
     conn->err = err;
@@ -143,7 +140,7 @@ static char *mongo_ns_to_cmd_db( const char *ns ) {
     int len = 0;
 
     for( current = (char *)ns; *current != '.'; current++ ) {
-      len++;
+        len++;
     }
 
     cmd_db_name = (char *)bson_malloc( len + 6 );
@@ -201,15 +198,15 @@ MONGO_EXPORT int mongo_validate_ns( mongo *conn, const char *ns ) {
     /* Go back and validate the database name. */
     for( current = (char *)db_name; *current != '.'; current++ ) {
         switch( *current ) {
-            case ' ':
-            case '$':
-            case '/':
-            case '\\':
-                __mongo_set_error( conn, MONGO_NS_INVALID,
-                    "Database name may not contain ' ', '$', '/', or '\\'", 0 );
-                return MONGO_ERROR;
-            default:
-                break;
+        case ' ':
+        case '$':
+        case '/':
+        case '\\':
+            __mongo_set_error( conn, MONGO_NS_INVALID,
+                               "Database name may not contain ' ', '$', '/', or '\\'", 0 );
+            return MONGO_ERROR;
+        default:
+            break;
         }
 
         ns_len++;
@@ -223,15 +220,15 @@ MONGO_EXPORT int mongo_validate_ns( mongo *conn, const char *ns ) {
 
         /* Cannot have two consecutive dots. */
         if( last && *last == '.' && *current == '.' ) {
-                __mongo_set_error( conn, MONGO_NS_INVALID,
-                    "Collection may not contain two consecutive '.'", 0 );
+            __mongo_set_error( conn, MONGO_NS_INVALID,
+                               "Collection may not contain two consecutive '.'", 0 );
             return MONGO_ERROR;
         }
 
         /* Cannot contain a '$' */
         if( *current == '$' ) {
             __mongo_set_error( conn, MONGO_NS_INVALID,
-                "Collection may not contain '$'", 0 );
+                               "Collection may not contain '$'", 0 );
             return MONGO_ERROR;
         }
 
@@ -249,7 +246,7 @@ MONGO_EXPORT int mongo_validate_ns( mongo *conn, const char *ns ) {
     /* Cannot end with a '.' */
     if( *(current - 1) == '.' ) {
         __mongo_set_error( conn, MONGO_NS_INVALID,
-            "Collection may not end with '.'", 0 );
+                           "Collection may not end with '.'", 0 );
         return MONGO_ERROR;
     }
 
@@ -376,10 +373,11 @@ static int mongo_check_is_master( mongo *conn ) {
         if( bson_find( &it, &out, "ismaster" ) )
             ismaster = bson_iterator_bool( &it );
         if( bson_find( &it, &out, "maxBsonObjectSize" ) ) {
-            max_bson_size = bson_iterator_int( &it ); 
+            max_bson_size = bson_iterator_int( &it );
         }
         conn->max_bson_size = max_bson_size;
-    } else {
+    }
+    else {
         return MONGO_ERROR;
     }
 
@@ -429,7 +427,7 @@ MONGO_EXPORT int mongo_client( mongo *conn , const char *host, int port ) {
 
 MONGO_EXPORT int mongo_connect( mongo *conn , const char *host, int port ) {
     int ret;
-    fprintf(stderr, "WARNING: mongo_connect() is deprecated, please use mongo_client()\n");
+    bson_errprintf("WARNING: mongo_connect() is deprecated, please use mongo_client()\n");
     ret = mongo_client( conn, host, port );
     mongo_set_write_concern( conn, 0 );
     return ret;
@@ -451,7 +449,7 @@ MONGO_EXPORT void mongo_replica_set_init( mongo *conn, const char *name ) {
 }
 
 MONGO_EXPORT void mongo_replset_init( mongo *conn, const char *name ) {
-    fprintf(stderr, "WARNING: mongo_replset_init() is deprecated, please use mongo_replica_set_init()\n");
+    bson_errprintf("WARNING: mongo_replset_init() is deprecated, please use mongo_replica_set_init()\n");
     mongo_replica_set_init( conn, name );
 }
 
@@ -489,7 +487,7 @@ MONGO_EXPORT void mongo_replica_set_add_seed( mongo *conn, const char *host, int
 }
 
 MONGO_EXPORT void mongo_replset_add_seed( mongo *conn, const char *host, int port ) {
-    fprintf(stderr, "WARNING: mongo_replset_add_seed() is deprecated, please use mongo_replica_set_add_seed()\n");
+    bson_errprintf("WARNING: mongo_replset_add_seed() is deprecated, please use mongo_replica_set_add_seed()\n");
     mongo_replica_set_add_node( &conn->replica_set->seeds, host, port );
 }
 
@@ -547,7 +545,7 @@ static void mongo_replica_set_check_seed( mongo *conn ) {
                 if( host_port ) {
                     mongo_parse_host( host_string, host_port );
                     mongo_replica_set_add_node( &conn->replica_set->hosts,
-                                            host_port->host, host_port->port );
+                                                host_port->host, host_port->port );
 
                     bson_free( host_port );
                     host_port = NULL;
@@ -582,7 +580,7 @@ static int mongo_replica_set_check_host( mongo *conn ) {
             ismaster = bson_iterator_bool( &it );
 
         if( bson_find( &it, &out, "maxBsonObjectSize" ) )
-            max_bson_size = bson_iterator_int( &it ); 
+            max_bson_size = bson_iterator_int( &it );
         conn->max_bson_size = max_bson_size;
 
         if( bson_find( &it, &out, "setName" ) ) {
@@ -599,7 +597,8 @@ static int mongo_replica_set_check_host( mongo *conn ) {
 
     if( ismaster ) {
         conn->replica_set->primary_connected = 1;
-    } else {
+    }
+    else {
         mongo_env_close_socket( conn->sock );
     }
 
@@ -632,7 +631,8 @@ MONGO_EXPORT int mongo_replica_set_client( mongo *conn ) {
     if( !conn->replica_set->hosts ) {
         conn->err = MONGO_CONN_NO_PRIMARY;
         return MONGO_ERROR;
-    } else {
+    }
+    else {
         node = conn->replica_set->hosts;
 
         while( node != NULL ) {
@@ -668,7 +668,7 @@ MONGO_EXPORT int mongo_replica_set_client( mongo *conn ) {
 
 MONGO_EXPORT int mongo_replset_connect( mongo *conn ) {
     int ret;
-    fprintf(stderr, "WARNING: mongo_replset_connect() is deprecated, please use mongo_replica_set_client()\n");
+    bson_errprintf("WARNING: mongo_replset_connect() is deprecated, please use mongo_replica_set_client()\n");
     ret = mongo_replica_set_client( conn );
     mongo_set_write_concern( conn, 0 );
     return ret;
@@ -692,7 +692,8 @@ MONGO_EXPORT int mongo_reconnect( mongo *conn ) {
         conn->replica_set->hosts = NULL;
         res = mongo_replica_set_client( conn );
         return res;
-    } else
+    }
+    else
         return mongo_env_socket_connect( conn, conn->primary->host, conn->primary->port );
 }
 
@@ -791,7 +792,7 @@ static int mongo_cursor_bson_valid( mongo_cursor *cursor, const bson *bson ) {
 }
 
 static int mongo_check_last_error( mongo *conn, const char *ns,
-    mongo_write_concern *write_concern ) {
+                                   mongo_write_concern *write_concern ) {
     int ret = MONGO_OK;
     bson response = {NULL, 0};
     bson fields;
@@ -806,10 +807,10 @@ static int mongo_check_last_error( mongo *conn, const char *ns,
         ret = MONGO_ERROR;
     else {
         if( ( bson_find( &it, &response, "$err" ) == BSON_STRING ) ||
-            ( bson_find( &it, &response, "err" ) == BSON_STRING ) ) {
+                ( bson_find( &it, &response, "err" ) == BSON_STRING ) ) {
 
             __mongo_set_error( conn, MONGO_WRITE_ERROR,
-                "See conn->lasterrstr for details.", 0 );
+                               "See conn->lasterrstr for details.", 0 );
             mongo_set_last_error( conn, &it, &response );
             ret = MONGO_ERROR;
         }
@@ -819,8 +820,8 @@ static int mongo_check_last_error( mongo *conn, const char *ns,
 }
 
 static int mongo_choose_write_concern( mongo *conn,
-    mongo_write_concern *custom_write_concern,
-    mongo_write_concern **write_concern ) {
+                                       mongo_write_concern *custom_write_concern,
+                                       mongo_write_concern **write_concern ) {
 
     if( custom_write_concern ) {
         *write_concern = custom_write_concern;
@@ -833,7 +834,7 @@ static int mongo_choose_write_concern( mongo *conn,
     }
     if( *write_concern && !((*write_concern)->cmd) ) {
         __mongo_set_error( conn, MONGO_WRITE_CONCERN_INVALID,
-            "Must call mongo_write_concern_finish() before using *write_concern.", 0 );
+                           "Must call mongo_write_concern_finish() before using *write_concern.", 0 );
         return MONGO_ERROR;
     }
     else
@@ -846,7 +847,7 @@ CRUD API
 **********************************************************************/
 
 MONGO_EXPORT int mongo_insert( mongo *conn, const char *ns,
-    const bson *bson, mongo_write_concern *custom_write_concern ) {
+                               const bson *bson, mongo_write_concern *custom_write_concern ) {
 
     char *data;
     mongo_message *mm;
@@ -860,7 +861,7 @@ MONGO_EXPORT int mongo_insert( mongo *conn, const char *ns,
     }
 
     if( mongo_choose_write_concern( conn, custom_write_concern,
-        &write_concern ) == MONGO_ERROR ) {
+                                    &write_concern ) == MONGO_ERROR ) {
         return MONGO_ERROR;
     }
 
@@ -891,8 +892,8 @@ MONGO_EXPORT int mongo_insert( mongo *conn, const char *ns,
 }
 
 MONGO_EXPORT int mongo_insert_batch( mongo *conn, const char *ns,
-    const bson **bsons, int count, mongo_write_concern *custom_write_concern,
-    int flags ) {
+                                     const bson **bsons, int count, mongo_write_concern *custom_write_concern,
+                                     int flags ) {
 
     mongo_message *mm;
     mongo_write_concern *write_concern = NULL;
@@ -916,7 +917,7 @@ MONGO_EXPORT int mongo_insert_batch( mongo *conn, const char *ns,
     }
 
     if( mongo_choose_write_concern( conn, custom_write_concern,
-        &write_concern ) == MONGO_ERROR ) {
+                                    &write_concern ) == MONGO_ERROR ) {
         return MONGO_ERROR;
     }
 
@@ -948,7 +949,7 @@ MONGO_EXPORT int mongo_insert_batch( mongo *conn, const char *ns,
 }
 
 MONGO_EXPORT int mongo_update( mongo *conn, const char *ns, const bson *cond,
-    const bson *op, int flags, mongo_write_concern *custom_write_concern ) {
+                               const bson *op, int flags, mongo_write_concern *custom_write_concern ) {
 
     char *data;
     mongo_message *mm;
@@ -962,7 +963,7 @@ MONGO_EXPORT int mongo_update( mongo *conn, const char *ns, const bson *cond,
     }
 
     if( mongo_choose_write_concern( conn, custom_write_concern,
-        &write_concern ) == MONGO_ERROR ) {
+                                    &write_concern ) == MONGO_ERROR ) {
         return MONGO_ERROR;
     }
 
@@ -996,7 +997,7 @@ MONGO_EXPORT int mongo_update( mongo *conn, const char *ns, const bson *cond,
 }
 
 MONGO_EXPORT int mongo_remove( mongo *conn, const char *ns, const bson *cond,
-    mongo_write_concern *custom_write_concern ) {
+                               mongo_write_concern *custom_write_concern ) {
 
     char *data;
     mongo_message *mm;
@@ -1010,16 +1011,16 @@ MONGO_EXPORT int mongo_remove( mongo *conn, const char *ns, const bson *cond,
     }
 
     if( mongo_choose_write_concern( conn, custom_write_concern,
-        &write_concern ) == MONGO_ERROR ) {
+                                    &write_concern ) == MONGO_ERROR ) {
         return MONGO_ERROR;
     }
 
     mm = mongo_message_create( 16  /* header */
-                              + 4  /* ZERO */
-                              + strlen( ns ) + 1
-                              + 4  /* ZERO */
-                              + bson_size( cond )
-                              , 0 , 0 , MONGO_OP_DELETE );
+                               + 4  /* ZERO */
+                               + strlen( ns ) + 1
+                               + 4  /* ZERO */
+                               + bson_size( cond )
+                               , 0 , 0 , MONGO_OP_DELETE );
 
     data = &mm->data;
     data = mongo_data_append32( data, &ZERO );
@@ -1109,7 +1110,7 @@ MONGO_EXPORT void mongo_write_concern_destroy( mongo_write_concern *write_concer
 }
 
 MONGO_EXPORT void mongo_set_write_concern( mongo *conn,
-    mongo_write_concern *write_concern ) {
+        mongo_write_concern *write_concern ) {
 
     conn->write_concern = write_concern;
 }
@@ -1191,13 +1192,16 @@ static int mongo_cursor_get_more( mongo_cursor *cursor ) {
     if( cursor->limit > 0 && cursor->seen >= cursor->limit ) {
         cursor->err = MONGO_CURSOR_EXHAUSTED;
         return MONGO_ERROR;
-    } else if( ! cursor->reply ) {
+    }
+    else if( ! cursor->reply ) {
         cursor->err = MONGO_CURSOR_INVALID;
         return MONGO_ERROR;
-    } else if( ! cursor->reply->fields.cursorID ) {
+    }
+    else if( ! cursor->reply->fields.cursorID ) {
         cursor->err = MONGO_CURSOR_EXHAUSTED;
         return MONGO_ERROR;
-    } else {
+    }
+    else {
         char *data;
         int sl = strlen( cursor->ns )+1;
         int limit = 0;
@@ -1238,7 +1242,7 @@ static int mongo_cursor_get_more( mongo_cursor *cursor ) {
 }
 
 MONGO_EXPORT mongo_cursor *mongo_find( mongo *conn, const char *ns, const bson *query,
-                          const bson *fields, int limit, int skip, int options ) {
+                                       const bson *fields, int limit, int skip, int options ) {
 
     mongo_cursor *cursor = ( mongo_cursor * )bson_malloc( sizeof( mongo_cursor ) );
     mongo_cursor_init( cursor, conn, ns );
@@ -1259,7 +1263,7 @@ MONGO_EXPORT mongo_cursor *mongo_find( mongo *conn, const char *ns, const bson *
 }
 
 MONGO_EXPORT int mongo_find_one( mongo *conn, const char *ns, const bson *query,
-                    const bson *fields, bson *out ) {
+                                 const bson *fields, bson *out ) {
 
     mongo_cursor cursor[1];
     mongo_cursor_init( cursor, conn, ns );
@@ -1271,12 +1275,13 @@ MONGO_EXPORT int mongo_find_one( mongo *conn, const char *ns, const bson *query,
         if( out ) {
             bson_init_size( out, bson_size( (bson *)&cursor->current ) );
             memcpy( out->data, cursor->current.data,
-                bson_size( (bson *)&cursor->current ) );
+                    bson_size( (bson *)&cursor->current ) );
             out->finished = 1;
         }
         mongo_cursor_destroy( cursor );
         return MONGO_OK;
-    } else {
+    }
+    else {
         mongo_cursor_destroy( cursor );
         return MONGO_ERROR;
     }
@@ -1368,7 +1373,8 @@ MONGO_EXPORT int mongo_cursor_next( mongo_cursor *cursor ) {
         }
 
         bson_init_finished_data( &cursor->current, &cursor->reply->objs );
-    } else {
+    }
+    else {
         bson_init_finished_data( &cursor->current, next_object );
     }
 
@@ -1458,7 +1464,7 @@ MONGO_EXPORT bson_bool_t mongo_create_simple_index( mongo *conn, const char *ns,
 }
 
 MONGO_EXPORT int mongo_create_capped_collection( mongo *conn, const char *db,
-    const char *collection, int size, int max, bson *out ) {
+        const char *collection, int size, int max, bson *out ) {
 
     bson b;
     int result;
@@ -1496,7 +1502,8 @@ MONGO_EXPORT double mongo_count( mongo *conn, const char *db, const char *ns, co
         bson_destroy( &cmd );
         bson_destroy( &out );
         return count;
-    } else {
+    }
+    else {
         bson_destroy( &out );
         bson_destroy( &cmd );
         return MONGO_ERROR;
@@ -1504,7 +1511,7 @@ MONGO_EXPORT double mongo_count( mongo *conn, const char *db, const char *ns, co
 }
 
 MONGO_EXPORT int mongo_run_command( mongo *conn, const char *db, const bson *command,
-                       bson *out ) {
+                                    bson *out ) {
     int ret = MONGO_OK;
     bson response = {NULL, 0};
     bson fields;
@@ -1529,7 +1536,8 @@ MONGO_EXPORT int mongo_run_command( mongo *conn, const char *db, const bson *com
             conn->err = MONGO_COMMAND_FAILED;
             bson_destroy( &response );
             ret = MONGO_ERROR;
-        } else {
+        }
+        else {
             if( out )
                 *out = response;
             else
@@ -1540,7 +1548,7 @@ MONGO_EXPORT int mongo_run_command( mongo *conn, const char *db, const bson *com
 }
 
 MONGO_EXPORT int mongo_simple_int_command( mongo *conn, const char *db,
-                              const char *cmdstr, int arg, bson *realout ) {
+        const char *cmdstr, int arg, bson *realout ) {
 
     bson out = {NULL, 0};
     bson cmd;
@@ -1563,7 +1571,7 @@ MONGO_EXPORT int mongo_simple_int_command( mongo *conn, const char *db,
 }
 
 MONGO_EXPORT int mongo_simple_str_command( mongo *conn, const char *db,
-                              const char *cmdstr, const char *arg, bson *realout ) {
+        const char *cmdstr, const char *arg, bson *realout ) {
 
     bson out = {NULL, 0};
     int result;
@@ -1717,7 +1725,8 @@ MONGO_EXPORT bson_bool_t mongo_cmd_authenticate( mongo *conn, const char *db, co
         bson_iterator it;
         bson_find( &it, &from_db, "nonce" );
         nonce = bson_iterator_string( &it );
-    } else {
+    }
+    else {
         return MONGO_ERROR;
     }
