+++ src/mongo.h
@@ -391,7 +391,7 @@ MONGO_EXPORT void mongo_destroy( mongo *conn );
  *
  */
 MONGO_EXPORT void mongo_set_write_concern( mongo *conn,
-    mongo_write_concern *write_concern );
+        mongo_write_concern *write_concern );
 
 
 /*********************************************************************
@@ -416,7 +416,7 @@ CRUD API
  *     on the bson struct for the reason.
  */
 MONGO_EXPORT int mongo_insert( mongo *conn, const char *ns, const bson *data,
-    mongo_write_concern *custom_write_concern );
+                               mongo_write_concern *custom_write_concern );
 
 /**
  * Insert a batch of BSON documents into a MongoDB server. This function
@@ -438,8 +438,8 @@ MONGO_EXPORT int mongo_insert( mongo *conn, const char *ns, const bson *data,
  *
  */
 MONGO_EXPORT int mongo_insert_batch( mongo *conn, const char *ns,
-    const bson **data, int num, mongo_write_concern *custom_write_concern,
-    int flags );
+                                     const bson **data, int num, mongo_write_concern *custom_write_concern,
+                                     int flags );
 
 /**
  * Update a document in a MongoDB server.
@@ -458,7 +458,7 @@ MONGO_EXPORT int mongo_insert_batch( mongo *conn, const char *ns,
  *
  */
 MONGO_EXPORT int mongo_update( mongo *conn, const char *ns, const bson *cond,
-    const bson *op, int flags, mongo_write_concern *custom_write_concern );
+                               const bson *op, int flags, mongo_write_concern *custom_write_concern );
 
 /**
  * Remove a document from a MongoDB server.
@@ -474,7 +474,7 @@ MONGO_EXPORT int mongo_update( mongo *conn, const char *ns, const bson *cond,
  * @return MONGO_OK or MONGO_ERROR with error stored in conn object.
  */
 MONGO_EXPORT int mongo_remove( mongo *conn, const char *ns, const bson *cond,
-    mongo_write_concern *custom_write_concern );
+                               mongo_write_concern *custom_write_concern );
 
 
 /*********************************************************************
@@ -522,7 +522,7 @@ Cursor API
  *     use the cursor builder API instead.
  */
 MONGO_EXPORT mongo_cursor *mongo_find( mongo *conn, const char *ns, const bson *query,
-                          const bson *fields, int limit, int skip, int options );
+                                       const bson *fields, int limit, int skip, int options );
 
 /**
  * Initalize a new cursor object.
@@ -631,7 +631,7 @@ MONGO_EXPORT int mongo_cursor_destroy( mongo_cursor *cursor );
  */
 /* out can be NULL if you don't care about results. useful for commands */
 MONGO_EXPORT int mongo_find_one( mongo *conn, const char *ns, const bson *query,
-                            const bson *fields, bson *out );
+                                 const bson *fields, bson *out );
 
 
 /*********************************************************************
@@ -650,7 +650,7 @@ Command API and Helpers
  *     MONGO_ERROR is returned.
  */
 MONGO_EXPORT double mongo_count( mongo *conn, const char *db, const char *coll,
-                     const bson *query );
+                                 const bson *query );
 
 /**
  * Create a compound index.
@@ -666,7 +666,7 @@ MONGO_EXPORT double mongo_count( mongo *conn, const char *db, const char *coll,
  * @return MONGO_OK if index is created successfully; otherwise, MONGO_ERROR.
  */
 MONGO_EXPORT int mongo_create_index( mongo *conn, const char *ns,
-    const bson *key, int options, bson *out );
+                                     const bson *key, int options, bson *out );
 
 /**
  * Create a capped collection.
@@ -681,7 +681,7 @@ MONGO_EXPORT int mongo_create_index( mongo *conn, const char *ns,
  *   number of documents.
  */
 MONGO_EXPORT int mongo_create_capped_collection( mongo *conn, const char *db,
-    const char *collection, int size, int max, bson *out );
+        const char *collection, int size, int max, bson *out );
 
 /**
  * Create an index with a single key.
@@ -695,7 +695,7 @@ MONGO_EXPORT int mongo_create_capped_collection( mongo *conn, const char *db,
  * @return true if the index was created.
  */
 MONGO_EXPORT bson_bool_t mongo_create_simple_index( mongo *conn, const char *ns,
-    const char *field, int options, bson *out );
+        const char *field, int options, bson *out );
 
 /**
  * Run a command on a MongoDB server.
@@ -708,7 +708,7 @@ MONGO_EXPORT bson_bool_t mongo_create_simple_index( mongo *conn, const char *ns,
  * @return MONGO_OK if the command ran without error.
  */
 MONGO_EXPORT int mongo_run_command( mongo *conn, const char *db,
-    const bson *command, bson *out );
+                                    const bson *command, bson *out );
 
 /**
  * Run a command that accepts a simple string key and integer value.
@@ -723,7 +723,7 @@ MONGO_EXPORT int mongo_run_command( mongo *conn, const char *db,
  *
  */
 MONGO_EXPORT int mongo_simple_int_command( mongo *conn, const char *db,
-                              const char *cmd, int arg, bson *out );
+        const char *cmd, int arg, bson *out );
 
 /**
  * Run a command that accepts a simple string key and value.
@@ -738,7 +738,7 @@ MONGO_EXPORT int mongo_simple_int_command( mongo *conn, const char *db,
  *
  */
 MONGO_EXPORT int mongo_simple_str_command( mongo *conn, const char *db,
-    const char *cmd, const char *arg, bson *out );
+        const char *cmd, const char *arg, bson *out );
 
 /**
  * Drop a database.
@@ -761,7 +761,7 @@ MONGO_EXPORT int mongo_cmd_drop_db( mongo *conn, const char *db );
  * @return true if the collection drop was successful.
  */
 MONGO_EXPORT int mongo_cmd_drop_collection( mongo *conn, const char *db,
-    const char *collection, bson *out );
+        const char *collection, bson *out );
 
 /**
  * Add a database user.
@@ -774,7 +774,7 @@ MONGO_EXPORT int mongo_cmd_drop_collection( mongo *conn, const char *db,
  * @return MONGO_OK or MONGO_ERROR.
   */
 MONGO_EXPORT int mongo_cmd_add_user( mongo *conn, const char *db,
-    const char *user, const char *pass );
+                                     const char *user, const char *pass );
 
 /**
  * Authenticate a user.
@@ -787,7 +787,7 @@ MONGO_EXPORT int mongo_cmd_add_user( mongo *conn, const char *db,
  * @return MONGO_OK on sucess and MONGO_ERROR on failure.
  */
 MONGO_EXPORT int mongo_cmd_authenticate( mongo *conn, const char *db,
-    const char *user, const char *pass );
+        const char *user, const char *pass );
 
 /**
  * Check if the current server is a master.
